pyproject.toml
[project]
name = "reed-data-distilation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
